refactor(metadata): one durable-handle store for both SQL dialects
sqlite and postgres each carried a full lock.DurableHandleStore: the same
eleven operations over the same 32 columns, differing only in placeholder
syntax and in how each expires a handle. Both now embed sql.DurableStore and
supply their statements through Dialect.Durable().

Only the expiry sweep stays per-dialect, because the two genuinely differ:
postgres deletes with interval arithmetic in one statement, while sqlite has no
interval arithmetic and modernc stores a time.Time in a layout SQLite's date
functions cannot parse, so it reads candidates and computes the deadline in Go.
Both now take their statements from the dialect rather than inlining them.

sqlite/durable_handles.go 408 -> 113 lines, postgres 418 -> 84, with the
duplicated scan helpers and column lists deleted.
